--- internal/sql/postgresql/column.go
@@ -1,6 +1,0 @@
-package postgresql
-
-/**
- @author: nizhenxian
- @date: 2023/10/30 21:26:37
-**/
