--- internal/sql/postgresql/table.go
@@ -1,9 +1,0 @@
-package postgresql
-
-type Table struct {
-	Name        string
-	IfNotExists bool
-	Columns     []*Column
-	Constraints []*Constraint
-	Comment     string
-}
